e2e: add threads, cores, dies and packages to topology
- Replace the numanodes variable with topology.
- Topology enables setting default distances between NUMA nodes
  on the same die, package, and on other packages. Defaults changed.
- Default CPU cores are hyperthreaded.
